requirements.txt
streamlit==1.37.1
transformers==4.43.3
torch>=2.2.0
python-dotenv==1.0.1
numpy==1.26.4
openai>=1.42.0
pillow
requests
sentence-transformers
scikit-learn
python-dotenv
